backend/requirements-sebou.txt
# Additional dependencies for Sebou satellite monitoring module (Phase 1)
earthengine-api==0.1.384
geopandas==0.14.1
rasterio==1.3.9
GDAL==3.8.0
GeoAlchemy2==0.14.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0
requests==2.31.0
Pillow==10.1.0
